+++ .idea/misc.xml
@@ -15,6 +15,7 @@
         <entry key="..\:/car-trackers/app/src/main/res/layout/activity_login.xml" value="0.5421875" />
         <entry key="..\:/car-trackers/app/src/main/res/layout/activity_main.xml" value="0.5421875" />
         <entry key="..\:/car-trackers/app/src/main/res/layout/activity_splash.xml" value="0.5421875" />
+        <entry key="..\:/car-trackers/app/src/main/res/layout/tracker_alert_layout.xml" value="0.5" />
         <entry key="..\:/car-trackers/app/src/main/res/layout/vello_alert_layout.xml" value="0.5421875" />
         <entry key="..\:/dev/trackers/app/src/main/res/layout/activity_main.xml" value="0.490036231884058" />
       </map>
