+++ app/src/main/java/com/cartrackers/app/features/splash/SplashActivity.kt
@@ -5,6 +5,7 @@ import android.graphics.Typeface
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.cartrackers.app.BuildConfig
+import com.cartrackers.app.comms.CarDialog
 import com.cartrackers.app.databinding.ActivitySplashBinding
 import com.cartrackers.app.di.providesSharedPrefStored
 import com.cartrackers.app.features.intro.IntroActivity
@@ -51,29 +52,14 @@ class SplashActivity: AppCompatActivity() {
     }
 
     private fun verifyConnection() {
-        val alertDialog = TrackerAlertDialog()
-        alertDialog.alertInitialize(
-            this,
-            "No Internet",
-            "Would you like to switch to offline?",
-            Typeface.SANS_SERIF,
-            Typeface.DEFAULT_BOLD,
-            isCancelable = true,
-            isNegativeBtnHide = false)
-        alertDialog.setPositive("YES", object : ListenerCallBack {
-            override fun onClick(dialog: TrackerAlertDialog) {
+        val value = CarDialog.build(this, "No Internet", "Would you like to switch to offline?")
+        when {
+            value -> {
                 saveInternetStatePref(true)
                 launchActivity()
-                dialog.dismiss()
             }
-        })
-        alertDialog.setNegative("NO", object : ListenerCallBack {
-            override fun onClick(dialog: TrackerAlertDialog) {
-                dialog.dismiss()
-                exitApp()
-            }
-        })
-        alertDialog.show()
+            else -> exitApp()
+        }
     }
 
     private fun exitApp() {
